services/hl-sage/requirements.txt
fastapi==0.114.2
uvicorn==0.30.1
nats-py==2.7.2
prometheus-client==0.20.0
pydantic==2.7.4
httpx==0.27.0
